Limit BAM files to 32 bits. XXX - it's not clear we ever use the size so we could drop this and make the limit be the 9.99GB that sfio has right now.
bk: 46acd0f6bf8Px6TYTz0JUoRHsgA4WA

+++ src/bam.c
@@ -493,7 +493,11 @@ bp_updateServer(char *range, char *list, int quiet)
 		out = fopen(tmpkeys, "w");
 		for (kv = mdbm_first(bp); kv.key.dsize; kv = mdbm_next(bp)) {
 			sprintf(buf, "BitKeeper/BAM/%s", kv.val.dptr);
-			fprintf(out, "|%u|%s\n", size(buf), kv.key.dptr);
+			/* If we want to allow >= 2^32 then fix sccs_delta() 
+			 * to not enforce the size limit.
+			 * And fix this to handle the %u portably.
+			 */
+			fprintf(out, "|%u|%s\n", (u32)size(buf), kv.key.dptr);
 		}
 		fclose(out);
 	} else {
